implement CodeParameter
