+++ phoebe/parameters/parameters.py
@@ -7891,8 +7891,6 @@ class StringParameter(Parameter):
         """
         self._readonly_check(**kwargs)
 
-        _orig_value = _deepcopy(value)
-
         try:
             value = str(value)
         except:
@@ -12651,3 +12649,48 @@ class JobParameter(Parameter):
         # TODO: options for whether to pass delete_volume
         self.crimpl_job.resubmit_script()
         self._value = 'unknown'
+
+
+class CodeParameter(StringParameter):
+    @staticmethod
+    def get_code_for_cls(cls, ignore=[]):
+        import inspect
+
+        # TODO: remove
+        from phoebe.features.component_features import Spot
+
+
+        code = inspect.getsource(cls.__init__)
+        for name, member in inspect.getmembers_static(cls):
+            if name in ignore:
+                continue
+            if name.startswith('__') and name not in ['__repr__']:
+                continue
+            if isinstance(member, property):
+                raise ValueError("input class cannot contain any properties (pass to ignore to skip during export)")
+            if hasattr(member, '__call__') or hasattr(member, '__wrapped__'):
+                code += inspect.getsource(member)
+            else:
+                value = getattr(cls, name)
+                if isinstance(value, str):
+                    value = f"'{value}'"
+                code += f"    {name} = {value}\n"
+        return code
+
+    @staticmethod
+    def get_class_from_code(code, clsname='ClassFromCode'):
+        if 'spot' in clsname.lower():
+            # TODO: do this in general with inspection
+            code = f"class {clsname}(Spot):\n    def __repr__(self):\n        return '<Feature {clsname}>'\n" + code
+        else:
+            code = f"class {clsname}:\n    def __repr__(self):\n        return '<Feature {clsname}>'\n" + code
+        namespace = {}
+        exec(code, globals(), namespace)
+        return namespace[clsname]
+
+    def set_value(self, value, **kwargs):
+        super().set_value(self.get_code_for_cls(value, ignore=['get_parameters']), **kwargs)
+
+    def get_value(self, **kwargs):
+        str_value = super().get_value(**kwargs)
+        return self.get_class_from_code(str_value, self.kind)
